# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Swish Stats — basketball stat capture API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games:
    get:
      operationId: listGames
      tags: [games]
      summary: List all games
      responses:
        "200":
          description: List of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    post:
      operationId: createGame
      tags: [games]
      summary: Create a new game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGameBody"
      responses:
        "201":
          description: Game created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  /games/{gameId}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get a single game
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          description: Game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateGame
      tags: [games]
      summary: Update a game
      parameters:
        - $ref: "#/components/parameters/GameId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGameBody"
      responses:
        "200":
          description: Updated game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteGame
      tags: [games]
      summary: Delete a game
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /games/{gameId}/box-score:
    get:
      operationId: getBoxScore
      tags: [aggregates]
      summary: Computed box score for a game
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          description: Box score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoxScore"
  /games/{gameId}/possessions:
    get:
      operationId: getPossessions
      tags: [aggregates]
      summary: Possession counts per team
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          description: Possession counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PossessionSummary"
  /games/{gameId}/clock:
    patch:
      operationId: updateClock
      tags: [games]
      summary: Update the live clock and current period
      parameters:
        - $ref: "#/components/parameters/GameId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClockBody"
      responses:
        "200":
          description: Updated game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
...
components:
  parameters:
...
    GameId:
      name: gameId
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    GameStatus:
      type: string
      enum: [setup, active, final]
    CaptureMode:
      type: string
      enum: [simple, complex]
    Game:
      type: object
      properties:
        id:
          type: integer
        leagueId:
          type: ["integer", "null"]
        competition:
          type: ["string", "null"]
        date:
          type: string
          format: date-time
        venue:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/GameStatus"
        captureMode:
          $ref: "#/components/schemas/CaptureMode"
        periodCount:
          type: integer
        periodDurationMins:
          type: integer
        overtimeCount:
          type: integer
        currentPeriod:
          type: integer
        clockSeconds:
          type: integer
        possessionTeamId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - date
        - status
        - captureMode
        - periodCount
        - periodDurationMins
        - overtimeCount
        - currentPeriod
        - clockSeconds
        - createdAt
        - updatedAt
    CreateGameBody:
      type: object
      properties:
        leagueId:
          type: ["integer", "null"]
        competition:
          type: ["string", "null"]
        date:
          type: string
          format: date-time
        venue:
          type: ["string", "null"]
        captureMode:
          $ref: "#/components/schemas/CaptureMode"
        periodCount:
          type: integer
        periodDurationMins:
          type: integer
    UpdateGameBody:
      type: object
      properties:
        competition:
          type: ["string", "null"]
        date:
          type: string
          format: date-time
        venue:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/GameStatus"
        captureMode:
          $ref: "#/components/schemas/CaptureMode"
        periodCount:
          type: integer
        periodDurationMins:
          type: integer
        overtimeCount:
          type: integer
        currentPeriod:
          type: integer
        clockSeconds:
          type: integer
        possessionTeamId:
          type: ["integer", "null"]
    UpdateClockBody:
      type: object
      properties:
        clockSeconds:
          type: integer
        currentPeriod:
          type: integer
...
    PlayerStatLine:
      type: object
      properties:
        playerId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        jerseyNumber:
          type: string
        teamId:
          type: integer
        points:
          type: integer
        fgMade:
          type: integer
        fgAttempted:
          type: integer
        twoPtMade:
          type: integer
        twoPtAttempted:
          type: integer
        threePtMade:
          type: integer
        threePtAttempted:
          type: integer
        ftMade:
          type: integer
        ftAttempted:
          type: integer
        offensiveRebounds:
          type: integer
        defensiveRebounds:
          type: integer
        totalRebounds:
          type: integer
        assists:
          type: integer
        steals:
          type: integer
        turnovers:
          type: integer
        blocks:
          type: integer
        personalFouls:
          type: integer
        technicalFouls:
          type: integer
        flagrantFouls:
          type: integer
      required:
        - playerId
        - firstName
        - lastName
        - jerseyNumber
        - teamId
        - points
        - fgMade
        - fgAttempted
        - twoPtMade
        - twoPtAttempted
        - threePtMade
        - threePtAttempted
        - ftMade
        - ftAttempted
        - offensiveRebounds
        - defensiveRebounds
        - totalRebounds
        - assists
        - steals
        - turnovers
        - blocks
        - personalFouls
        - technicalFouls
        - flagrantFouls
    TeamBoxScore:
      type: object
      properties:
        teamId:
          type: integer
        name:
          type: string
        abbreviation:
          type: string
        isHome:
          type: boolean
        totalPoints:
          type: integer
        possessions:
          type: integer
        players:
          type: array
          items:
            $ref: "#/components/schemas/PlayerStatLine"
        teamTotals:
          $ref: "#/components/schemas/PlayerStatLine"
      required:
        - teamId
        - name
        - abbreviation
        - isHome
        - totalPoints
        - possessions
        - players
        - teamTotals
    BoxScore:
      type: object
      properties:
        gameId:
          type: integer
        home:
          $ref: "#/components/schemas/TeamBoxScore"
        away:
          $ref: "#/components/schemas/TeamBoxScore"
      required: [gameId, home, away]
    PossessionSummary:
      type: object
      properties:
        gameId:
          type: integer
        currentPossessionTeamId:
          type: ["integer", "null"]
        homeTeamId:
          type: integer
        awayTeamId:
          type: integer
        homePossessions:
          type: integer
        awayPossessions:
          type: integer
      required:
        - gameId
        - homeTeamId
        - awayTeamId
        - homePossessions
        - awayPossessions